Rearrange calls and logic

+++ services/alp-dataflow/src/prefect/prefect.service.ts
@@ -203,16 +203,6 @@ export class PrefectService {
     let flowMetadataInput
     let existingFlowMetadata
     try {
-      // prepare metadata input
-      if (defaultDeploymentInfo && defaultDeploymentInfo.url) {
-        flowMetadataInput = await this.prepareFlowMetadata(deploymentFolderPath, defaultDeploymentInfo.url)
-      } else {
-        flowMetadataInput = await this.prepareFlowMetadata(deploymentFolderPath, url)
-      }
-
-      // if metadata with same flowId exists
-      existingFlowMetadata = await this.prefectFlowService.getFlowMetadataById(flowMetadataInput.flowId)
-
       if (defaultPluginId) {
         await this.prefectFlowService.updateDefaultPluginStatus(defaultPluginId, PluginUploadStatus.INSTALLING)
       }
@@ -235,14 +225,21 @@ export class PrefectService {
       }
 
       await this.prefectExecutionClient.executePipInstall(userId, modifiedFileStem)
-      if (existingFlowMetadata) {
-        await this.prefectFlowService.deleteFlowMetadata(existingFlowMetadata.flowId)
+
+      if (defaultDeploymentInfo && defaultDeploymentInfo.url) {
+        flowMetadataInput = await this.prepareFlowMetadata(deploymentFolderPath, defaultDeploymentInfo.url)
+      } else {
+        flowMetadataInput = await this.prepareFlowMetadata(deploymentFolderPath, url)
       }
+
       await this.prefectFlowService.createFlowMetadata(flowMetadataInput)
     } catch (err) {
-      if (!existingFlowMetadata) {
-        await this.prefectFlowService.deleteFlowMetadata(flowMetadataInput.flowId)
-        await this.prefectApi.deleteFlow(flowMetadataInput.flowId)
+      if (flowMetadataInput) {
+        existingFlowMetadata = await this.prefectFlowService.getFlowMetadataById(flowMetadataInput.flowId)
+        if (!existingFlowMetadata) {
+          await this.prefectFlowService.deleteFlowMetadata(flowMetadataInput.flowId)
+          await this.prefectApi.deleteFlow(flowMetadataInput.flowId)
+        }
       }
       this.deleteDeploymentFolder(deploymentFolderPath)
       const errorMessage = `Error installing pip package, check if package is valid`
@@ -269,6 +266,7 @@ export class PrefectService {
         await this.prefectFlowService.updateDefaultPluginStatus(defaultPluginId, PluginUploadStatus.COMPLETE)
       }
     } catch (err) {
+      existingFlowMetadata = await this.prefectFlowService.getFlowMetadataById(flowMetadataInput.flowId)
       if (!existingFlowMetadata) {
         await this.prefectFlowService.deleteFlowMetadata(flowMetadataInput.flowId)
         await this.prefectApi.deleteFlow(flowMetadataInput.flowId)
@@ -340,16 +338,6 @@ export class PrefectService {
     let flowMetadataInput
     let existingFlowMetadata
     try {
-      // prepare metadata input
-      if (defaultDeploymentInfo && defaultDeploymentInfo.url) {
-        flowMetadataInput = await this.prepareFlowMetadata(deploymentFolderPath, defaultDeploymentInfo.url)
-      } else {
-        flowMetadataInput = await this.prepareFlowMetadata(deploymentFolderPath, url)
-      }
-
-      // if metadata with same flowId exists
-      existingFlowMetadata = await this.prefectFlowService.getFlowMetadataById(flowMetadataInput.flowId)
-
       defaultDeploymentInfo.defaultPluginId
       if (defaultPluginId) {
         await this.prefectFlowService.updateDefaultPluginStatus(defaultPluginId, PluginUploadStatus.INSTALLING)
@@ -373,13 +361,21 @@ export class PrefectService {
       }
 
       await this.prefectExecutionClient.executePipInstall(userId, modifiedFileStem)
-      if (existingFlowMetadata) {
-        await this.prefectFlowService.deleteFlowMetadata(existingFlowMetadata.flowId)
+
+      if (defaultDeploymentInfo && defaultDeploymentInfo.url) {
+        flowMetadataInput = await this.prepareFlowMetadata(deploymentFolderPath, defaultDeploymentInfo.url)
+      } else {
+        flowMetadataInput = await this.prepareFlowMetadata(deploymentFolderPath, url)
       }
+
+      await this.prefectFlowService.createFlowMetadata(flowMetadataInput)
     } catch (err) {
-      if (!existingFlowMetadata) {
-        await this.prefectFlowService.deleteFlowMetadata(flowMetadataInput.flowId)
-        await this.prefectApi.deleteFlow(flowMetadataInput.flowId)
+      if (flowMetadataInput) {
+        existingFlowMetadata = await this.prefectFlowService.getFlowMetadataById(flowMetadataInput.flowId)
+        if (!existingFlowMetadata) {
+          await this.prefectFlowService.deleteFlowMetadata(flowMetadataInput.flowId)
+          await this.prefectApi.deleteFlow(flowMetadataInput.flowId)
+        }
       }
       this.deleteDeploymentFolder(deploymentFolderPath)
       const errorMessage = `Error installing pip package, check if package is valid`
@@ -405,6 +401,7 @@ export class PrefectService {
         await this.prefectFlowService.updateDefaultPluginStatus(defaultPluginId, PluginUploadStatus.COMPLETE)
       }
     } catch (err) {
+      existingFlowMetadata = await this.prefectFlowService.getFlowMetadataById(flowMetadataInput.flowId)
       if (!existingFlowMetadata) {
         await this.prefectFlowService.deleteFlowMetadata(flowMetadataInput.flowId)
         await this.prefectApi.deleteFlow(flowMetadataInput.flowId)
